requirements.txt
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
apprise==1.9.1
asgi-lifespan==2.1.0
asttokens==3.0.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.3.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
coolname==2.2.0
croniter==5.0.1
cryptography==44.0.0
dateparser==1.2.0
debugpy==1.8.11
decorator==5.1.1
Deprecated==1.2.15
dlt==1.5.0
docker==7.1.0
duckdb==1.1.3
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
giturlparse==0.12.0
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
graphviz==0.20.3
greenlet==3.1.1
griffe==1.5.4
h11==0.14.0
h2==4.1.0
hexbytes==1.2.1
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
humanize==4.11.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jinja2-humanize-extension==0.4.0
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
makefun==1.15.6
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.2.1
oauthlib==3.2.2
opentelemetry-api==1.29.0
orjson==3.10.13
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pathvalidate==3.2.3
pendulum==3.0.0
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
prefect==3.1.11
prometheus_client==0.21.1
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic-extra-types==2.10.1
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-socks==2.6.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
readchar==4.2.1
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requirements-parser==0.11.0
rfc3339-validator==0.1.4
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
semver==3.0.2
shellingham==1.5.4
simplejson==3.19.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
tenacity==9.0.0
text-unidecode==1.3
time-machine==2.16.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
traitlets==5.14.3
typer==0.13.1
types-setuptools==75.6.0.20241223
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==13.1
wrapt==1.17.0
zipp==3.21.0
